+++ src/api/http/api.go
@@ -541,7 +541,8 @@ func (self *HttpServer) tryAsClusterAdmin(w libhttp.ResponseWriter, r *libhttp.R
 }
 
 type NewUser struct {
-	Name     string `json:"username"`
+	Username string `json:"username"`
+	Name     string `json:"name"`
 	Password string `json:"password"`
 }
 
@@ -589,11 +590,15 @@ func (self *HttpServer) createClusterAdmin(w libhttp.ResponseWriter, r *libhttp.
 	}
 
 	self.tryAsClusterAdmin(w, r, func(u common.User) (int, interface{}) {
-		if err := self.userManager.CreateClusterAdminUser(u, newUser.Name); err != nil {
+		username := newUser.Name
+		if username == "" {
+			username = newUser.Username
+		}
+		if err := self.userManager.CreateClusterAdminUser(u, username); err != nil {
 			errorStr := err.Error()
 			return errorToStatusCode(err), errorStr
 		}
-		if err := self.userManager.ChangeClusterAdminPassword(u, newUser.Name, newUser.Password); err != nil {
+		if err := self.userManager.ChangeClusterAdminPassword(u, username, newUser.Password); err != nil {
 			return errorToStatusCode(err), err.Error()
 		}
 		return libhttp.StatusOK, nil
@@ -725,10 +730,14 @@ func (self *HttpServer) createDbUser(w libhttp.ResponseWriter, r *libhttp.Reques
 	db := r.URL.Query().Get(":db")
 
 	self.tryAsDbUserAndClusterAdmin(w, r, func(u common.User) (int, interface{}) {
-		if err := self.userManager.CreateDbUser(u, db, newUser.Name); err != nil {
+		username := newUser.Name
+		if username == "" {
+			username = newUser.Username
+		}
+		if err := self.userManager.CreateDbUser(u, db, username); err != nil {
 			return errorToStatusCode(err), err.Error()
 		}
-		if err := self.userManager.ChangeDbUserPassword(u, db, newUser.Name, newUser.Password); err != nil {
+		if err := self.userManager.ChangeDbUserPassword(u, db, username, newUser.Password); err != nil {
 			return libhttp.StatusUnauthorized, err.Error()
 		}
 		return libhttp.StatusOK, nil